Move SELECT analysis code into separate modules
