+++ src/sqlfluff/core/rules/analysis/__init__.py
@@ -1,0 +1,2 @@
+from sqlfluff.core.rules.analysis.select import get_aliases_from_select, get_select_statement_info
+from sqlfluff.core.rules.analysis.select_crawler import SelectCrawler
